[UPDATE] Square root- Newton Method - correct data types

+++ Mathematical/NewtonMethodSqrt/main.cpp
@@ -1,17 +1,17 @@
 #include <iostream>
 #include <cmath>
 
-double newton_sqrt(double n, double tolerance=0.0000001);
+long double newton_sqrt(long double n, long double tolerance=0.0000001);
 
 int main() {
-	double n = 18446744073709551615.0;
-	double root = newton_sqrt(n);
+	long double n = 18446744073709551615.0;
+	long double root = newton_sqrt(n);
 	std::cout << root << '\n' << std::sqrt(n) << '\n' << std::abs(root - std::sqrt(n));
 	return 0;
 }
 
-double newton_sqrt(double n, double tolerance) {
-	double a = n / 2;
+long double newton_sqrt(long double n, long double tolerance) {
+	long double a = n / 2;
 	while(std::abs((n / a) -a) > tolerance) {
 		a = (a + (n/a))/2;
 	}
